Add dark/light theme functionality
Changes the scribe buttons to the right color.
Implement a functionality for toggling between themes, as well as detecting which theme the system is using during startup, however, it can not detect changes during runtime as of now. The scribe logo changes depending on light/dark mode, and now has the right color.